+++ packages/camera/camera/ios/Classes/CameraPlugin.m
@@ -162,6 +162,17 @@ static FlashMode getFlashModeForString(NSString *mode) {
   }
 }
 
+static OSType getVideoFormatFromString(NSString *videoFormatString) {
+  if ([videoFormatString isEqualToString:@"bgra8888"]) {
+    return kCVPixelFormatType_32BGRA;
+  } else if ([videoFormatString isEqualToString:@"yuv420"]) {
+    return kCVPixelFormatType_420YpCbCr8BiPlanarVideoRange;
+  } else {
+    NSLog(@"The selected imageFormatGroup is not supported by iOS. Defaulting to brga8888");
+    return kCVPixelFormatType_32BGRA;
+  }
+}
+
 static AVCaptureFlashMode getAVCaptureFlashModeForFlashMode(FlashMode mode) {
   switch (mode) {
     case FlashModeOff:
@@ -173,6 +184,83 @@ static AVCaptureFlashMode getAVCaptureFlashModeForFlashMode(FlashMode mode) {
     case FlashModeTorch:
     default:
       return -1;
+  }
+}
+
+// Mirrors ExposureMode in camera.dart
+typedef enum {
+  ExposureModeAuto,
+  ExposureModeLocked,
+
+} ExposureMode;
+
+static NSString *getStringForExposureMode(ExposureMode mode) {
+  switch (mode) {
+    case ExposureModeAuto:
+      return @"auto";
+    case ExposureModeLocked:
+      return @"locked";
+  }
+  NSError *error = [NSError errorWithDomain:NSCocoaErrorDomain
+                                       code:NSURLErrorUnknown
+                                   userInfo:@{
+                                     NSLocalizedDescriptionKey : [NSString
+                                         stringWithFormat:@"Unknown string for exposure mode"]
+                                   }];
+  @throw error;
+}
+
+static ExposureMode getExposureModeForString(NSString *mode) {
+  if ([mode isEqualToString:@"auto"]) {
+    return ExposureModeAuto;
+  } else if ([mode isEqualToString:@"locked"]) {
+    return ExposureModeLocked;
+  } else {
+    NSError *error = [NSError errorWithDomain:NSCocoaErrorDomain
+                                         code:NSURLErrorUnknown
+                                     userInfo:@{
+                                       NSLocalizedDescriptionKey : [NSString
+                                           stringWithFormat:@"Unknown exposure mode %@", mode]
+                                     }];
+    @throw error;
+  }
+}
+
+// Mirrors FocusMode in camera.dart
+typedef enum {
+  FocusModeAuto,
+  FocusModeLocked,
+} FocusMode;
+
+static NSString *getStringForFocusMode(FocusMode mode) {
+  switch (mode) {
+    case FocusModeAuto:
+      return @"auto";
+    case FocusModeLocked:
+      return @"locked";
+  }
+  NSError *error = [NSError errorWithDomain:NSCocoaErrorDomain
+                                       code:NSURLErrorUnknown
+                                   userInfo:@{
+                                     NSLocalizedDescriptionKey : [NSString
+                                         stringWithFormat:@"Unknown string for focus mode"]
+                                   }];
+  @throw error;
+}
+
+static FocusMode getFocusModeForString(NSString *mode) {
+  if ([mode isEqualToString:@"auto"]) {
+    return FocusModeAuto;
+  } else if ([mode isEqualToString:@"locked"]) {
+    return FocusModeLocked;
+  } else {
+    NSError *error = [NSError errorWithDomain:NSCocoaErrorDomain
+                                         code:NSURLErrorUnknown
+                                     userInfo:@{
+                                       NSLocalizedDescriptionKey : [NSString
+                                           stringWithFormat:@"Unknown focus mode %@", mode]
+                                     }];
+    @throw error;
   }
 }
 
@@ -243,6 +331,8 @@ static ResolutionPreset getResolutionPresetForString(NSString *preset) {
 @property(assign, nonatomic) BOOL isAudioSetup;
 @property(assign, nonatomic) BOOL isStreamingImages;
 @property(assign, nonatomic) ResolutionPreset resolutionPreset;
+@property(assign, nonatomic) ExposureMode exposureMode;
+@property(assign, nonatomic) FocusMode focusMode;
 @property(assign, nonatomic) FlashMode flashMode;
 @property(assign, nonatomic) CMTime lastVideoSampleTime;
 @property(assign, nonatomic) CMTime lastAudioSampleTime;
@@ -256,7 +346,7 @@ static ResolutionPreset getResolutionPresetForString(NSString *preset) {
   dispatch_queue_t _dispatchQueue;
 }
 // Format used for video and image streaming.
-FourCharCode const videoFormat = kCVPixelFormatType_32BGRA;
+FourCharCode videoFormat = kCVPixelFormatType_32BGRA;
 NSString *const errorMethod = @"error";
 
 - (instancetype)initWithCameraName:(NSString *)cameraName
@@ -275,11 +365,14 @@ NSString *const errorMethod = @"error";
   _dispatchQueue = dispatchQueue;
   _captureSession = [[AVCaptureSession alloc] init];
   _flashMode = FlashModeAuto;
+  _exposureMode = ExposureModeAuto;
+  _focusMode = FocusModeAuto;
 
   _captureDevice = [AVCaptureDevice deviceWithUniqueID:cameraName];
   NSError *localError = nil;
   _captureVideoInput = [AVCaptureDeviceInput deviceInputWithDevice:_captureDevice
                                                              error:&localError];
+
   if (localError) {
     *error = localError;
     return nil;
@@ -340,6 +433,7 @@ NSString *const errorMethod = @"error";
     result(getFlutterError(error));
     return;
   }
+
   [_capturePhotoOutput
       capturePhotoWithSettings:settings
                       delegate:[[FLTSavePhotoDelegate alloc] initWithPath:path
@@ -771,6 +865,103 @@ NSString *const errorMethod = @"error";
   result(nil);
 }
 
+- (void)setExposureModeWithResult:(FlutterResult)result mode:(NSString *)modeStr {
+  ExposureMode mode;
+  @try {
+    mode = getExposureModeForString(modeStr);
+  } @catch (NSError *e) {
+    result(getFlutterError(e));
+    return;
+  }
+  _exposureMode = mode;
+  [self applyExposureMode];
+  result(nil);
+}
+
+- (void)applyExposureMode {
+  [_captureDevice lockForConfiguration:nil];
+  switch (_exposureMode) {
+    case ExposureModeLocked:
+      [_captureDevice setExposureMode:AVCaptureExposureModeAutoExpose];
+      break;
+    case ExposureModeAuto:
+      if ([_captureDevice isExposureModeSupported:AVCaptureExposureModeContinuousAutoExposure]) {
+        [_captureDevice setExposureMode:AVCaptureExposureModeContinuousAutoExposure];
+      } else {
+        [_captureDevice setExposureMode:AVCaptureExposureModeAutoExpose];
+      }
+      break;
+  }
+  [_captureDevice unlockForConfiguration];
+}
+
+- (void)setFocusModeWithResult:(FlutterResult)result mode:(NSString *)modeStr {
+  FocusMode mode;
+  @try {
+    mode = getFocusModeForString(modeStr);
+  } @catch (NSError *e) {
+    result(getFlutterError(e));
+    return;
+  }
+  _focusMode = mode;
+  [self applyFocusMode];
+  result(nil);
+}
+
+- (void)applyFocusMode {
+  [_captureDevice lockForConfiguration:nil];
+  switch (_focusMode) {
+    case FocusModeLocked:
+      [_captureDevice setFocusMode:AVCaptureFocusModeAutoFocus];
+      break;
+    case FocusModeAuto:
+      if ([_captureDevice isFocusModeSupported:AVCaptureFocusModeContinuousAutoFocus]) {
+        [_captureDevice setFocusMode:AVCaptureFocusModeContinuousAutoFocus];
+      } else {
+        [_captureDevice setFocusMode:AVCaptureFocusModeAutoFocus];
+      }
+      break;
+  }
+  [_captureDevice unlockForConfiguration];
+}
+
+- (void)setExposurePointWithResult:(FlutterResult)result x:(double)x y:(double)y {
+  if (!_captureDevice.isExposurePointOfInterestSupported) {
+    result([FlutterError errorWithCode:@"setExposurePointFailed"
+                               message:@"Device does not have exposure point capabilities"
+                               details:nil]);
+    return;
+  }
+  [_captureDevice lockForConfiguration:nil];
+  [_captureDevice setExposurePointOfInterest:CGPointMake(y, 1 - x)];
+  [_captureDevice unlockForConfiguration];
+  // Retrigger auto exposure
+  [self applyExposureMode];
+  result(nil);
+}
+
+- (void)setFocusPointWithResult:(FlutterResult)result x:(double)x y:(double)y {
+  if (!_captureDevice.isFocusPointOfInterestSupported) {
+    result([FlutterError errorWithCode:@"setFocusPointFailed"
+                               message:@"Device does not have focus point capabilities"
+                               details:nil]);
+    return;
+  }
+  [_captureDevice lockForConfiguration:nil];
+  [_captureDevice setFocusPointOfInterest:CGPointMake(y, 1 - x)];
+  [_captureDevice unlockForConfiguration];
+  // Retrigger auto focus
+  [self applyFocusMode];
+  result(nil);
+}
+
+- (void)setExposureOffsetWithResult:(FlutterResult)result offset:(double)offset {
+  [_captureDevice lockForConfiguration:nil];
+  [_captureDevice setExposureTargetBias:offset completionHandler:nil];
+  [_captureDevice unlockForConfiguration];
+  result(@(offset));
+}
+
 - (void)startImageStreamWithMessenger:(NSObject<FlutterBinaryMessenger> *)messenger {
   if (!_isStreamingImages) {
     FlutterEventChannel *eventChannel =
@@ -1051,6 +1242,9 @@ NSString *const errorMethod = @"error";
     NSDictionary *argsMap = call.arguments;
     NSUInteger cameraId = ((NSNumber *)argsMap[@"cameraId"]).unsignedIntegerValue;
     if ([@"initialize" isEqualToString:call.method]) {
+      NSString *videoFormatValue = ((NSString *)argsMap[@"imageFormatGroup"]);
+      videoFormat = getVideoFormatFromString(videoFormatValue);
+
       __weak CameraPlugin *weakSelf = self;
       _camera.onFrameAvailable = ^{
         [weakSelf.registry textureFrameAvailable:cameraId];
@@ -1060,11 +1254,17 @@ NSString *const errorMethod = @"error";
                                                            (unsigned long)cameraId]
                 binaryMessenger:_messenger];
       _camera.methodChannel = methodChannel;
-      [methodChannel invokeMethod:@"initialized"
-                        arguments:@{
-                          @"previewWidth" : @(_camera.previewSize.width),
-                          @"previewHeight" : @(_camera.previewSize.height)
-                        }];
+      [methodChannel
+          invokeMethod:@"initialized"
+             arguments:@{
+               @"previewWidth" : @(_camera.previewSize.width),
+               @"previewHeight" : @(_camera.previewSize.height),
+               @"exposureMode" : getStringForExposureMode([_camera exposureMode]),
+               @"focusMode" : getStringForFocusMode([_camera focusMode]),
+               @"exposurePointSupported" :
+                   @([_camera.captureDevice isExposurePointOfInterestSupported]),
+               @"focusPointSupported" : @([_camera.captureDevice isFocusPointOfInterestSupported]),
+             }];
       [_camera start];
       result(nil);
     } else if ([@"takePicture" isEqualToString:call.method]) {
@@ -1098,6 +1298,37 @@ NSString *const errorMethod = @"error";
       [_camera setZoomLevel:zoom Result:result];
     } else if ([@"setFlashMode" isEqualToString:call.method]) {
       [_camera setFlashModeWithResult:result mode:call.arguments[@"mode"]];
+    } else if ([@"setExposureMode" isEqualToString:call.method]) {
+      [_camera setExposureModeWithResult:result mode:call.arguments[@"mode"]];
+    } else if ([@"setExposurePoint" isEqualToString:call.method]) {
+      BOOL reset = ((NSNumber *)call.arguments[@"reset"]).boolValue;
+      double x = 0.5;
+      double y = 0.5;
+      if (!reset) {
+        x = ((NSNumber *)call.arguments[@"x"]).doubleValue;
+        y = ((NSNumber *)call.arguments[@"y"]).doubleValue;
+      }
+      [_camera setExposurePointWithResult:result x:x y:y];
+    } else if ([@"getMinExposureOffset" isEqualToString:call.method]) {
+      result(@(_camera.captureDevice.minExposureTargetBias));
+    } else if ([@"getMaxExposureOffset" isEqualToString:call.method]) {
+      result(@(_camera.captureDevice.maxExposureTargetBias));
+    } else if ([@"getExposureOffsetStepSize" isEqualToString:call.method]) {
+      result(@(0.0));
+    } else if ([@"setExposureOffset" isEqualToString:call.method]) {
+      [_camera setExposureOffsetWithResult:result
+                                    offset:((NSNumber *)call.arguments[@"offset"]).doubleValue];
+    } else if ([@"setFocusMode" isEqualToString:call.method]) {
+      [_camera setFocusModeWithResult:result mode:call.arguments[@"mode"]];
+    } else if ([@"setFocusPoint" isEqualToString:call.method]) {
+      BOOL reset = ((NSNumber *)call.arguments[@"reset"]).boolValue;
+      double x = 0.5;
+      double y = 0.5;
+      if (!reset) {
+        x = ((NSNumber *)call.arguments[@"x"]).doubleValue;
+        y = ((NSNumber *)call.arguments[@"y"]).doubleValue;
+      }
+      [_camera setFocusPointWithResult:result x:x y:y];
     } else {
       result(FlutterMethodNotImplemented);
     }